+++ src/main/java/woowacourse/shoppingcart/dto/response/EmailDuplicationResponse.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.response;
 
 public class EmailDuplicationResponse {
     private boolean isDuplicated;
