+++ src/main/java/woowacourse/shoppingcart/dto/response/ProductExistingInCartResponse.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.response;
 
 public class ProductExistingInCartResponse {
     private final boolean exists;
