+++ src/test/java/woowacourse/Fixtures.java
@@ -5,8 +5,8 @@ import java.time.format.DateTimeFormatter;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import woowacourse.auth.dto.TokenRequest;
-import woowacourse.shoppingcart.dto.AddressRequest;
-import woowacourse.shoppingcart.dto.CustomerRequest;
+import woowacourse.shoppingcart.dto.request.AddressRequest;
+import woowacourse.shoppingcart.dto.request.CustomerRequest;
 import woowacourse.shoppingcart.entity.AddressEntity;
 import woowacourse.shoppingcart.entity.CustomerEntity;
 import woowacourse.shoppingcart.entity.PrivacyEntity;
